chapter_4/grpc_project/pyproject.toml
[project]
name = "grpc-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
